Move repository functionality from client to service
