+++ .mypy.ini
@@ -58,7 +58,8 @@ modules =
     lambdas.indexer.app,
     azul.schemas,
     azul.indexer.index_controller,
-    azul.indexer.index_queue_service
+    azul.indexer.index_queue_service,
+    azul.indexer.index_repository_service
 packages =
     azul.openapi
